ARCH-1: extract task dispatch arm into commands::tasks::run (iter-225)
